+++ CHANGELOG.md
@@ -7,6 +7,10 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Changed
+* `providers.TypeORMProvider`'s `entityManager` is now `readonly`
+* TypeORM dependency has been updated to `0.1.1` release
+
 ## [0.5.1] - 2017-10-06
 
 ### Fixed